requirements.txt
pygame>=2.0.0
pycairo>=1.20.0